Split most Plugins into Editor and Runtime assemblies
