+++ Packages/com.unity.inputsystem/Editor/Controls.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dce36297a61ee468f92d0fe7f7e28441
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
