+++ Packages/com.unity.inputsystem/Editor/Plugins.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9264bbaac8fbe403b987d9584bb1ffc5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
