+++ Packages/com.unity.inputsystem/Editor/Plugins/EnhancedTouch.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 36cf437a9315840c0a1df7d24d3d2787
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
