+++ Packages/com.unity.inputsystem/Editor/Plugins/HID.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e38cdda355b6f4f8fb66876f964f570e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
