+++ Packages/com.unity.inputsystem/Editor/Plugins/HID/HIDSupportEditorInitializer.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: efc4aca331ba242fab026845289179ff
