+++ Packages/com.unity.inputsystem/Editor/Plugins/InputForUI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e37b431b8b93e4719910a045d05ba3cb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
