+++ Packages/com.unity.inputsystem/Editor/Plugins/OnScreen.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f027141a193544959a71baea4a57d3f7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
