+++ Packages/com.unity.inputsystem/Editor/Plugins/OnScreen/OnScreenStickEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 839e0066c5f854ddda076a4ff2f80a7b
