+++ Packages/com.unity.inputsystem/Editor/Plugins/OnScreen/UGUIOnScreenControlEditorUtils.cs
@@ -1,0 +1,13 @@
+using UnityEditor;
+
+namespace UnityEngine.InputSystem.OnScreen
+{
+    internal static class UGUIOnScreenControlEditorUtils
+    {
+        public static void ShowWarningIfNotPartOfCanvasHierarchy(OnScreenControl target)
+        {
+            if (UGUIOnScreenControlUtils.GetCanvasRectTransform(target.transform) == null)
+                UnityEditor.EditorGUILayout.HelpBox(target.GetWarningMessage(), UnityEditor.MessageType.Warning);
+        }
+    }
+}
