+++ Packages/com.unity.inputsystem/Editor/Plugins/OnScreen/UGUIOnScreenControlEditorUtils.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 91a3b2b6adad844ed9b680b378d8485a
