+++ Packages/com.unity.inputsystem/Editor/Plugins/PlayerInput.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cf0897af2cae4ebbaf8b79eaae98fb08
+guid: 95a3b96623aff409fad313b971f84743
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
