+++ Packages/com.unity.inputsystem/Editor/Plugins/Steam.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0a40837a6cb5749b69436d6839711335
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
