+++ Packages/com.unity.inputsystem/Editor/Plugins/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 03672f9d7e73f4134820c35a3cd46a86
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
