+++ Packages/com.unity.inputsystem/Editor/Plugins/UnityRemote.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b67ddb076aa7b4301b132a1012ba666d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
