+++ Packages/com.unity.inputsystem/Runtime/Plugins/OnScreen/OnScreenButton.cs
@@ -2,10 +2,6 @@
 using UnityEngine.EventSystems;
 using UnityEngine.InputSystem.Layouts;
 
-#if UNITY_EDITOR
-using UnityEngine.InputSystem.Editor;
-#endif
-
 ////TODO: custom icon for OnScreenButton component
 
 namespace UnityEngine.InputSystem.OnScreen
@@ -41,41 +37,13 @@ namespace UnityEngine.InputSystem.OnScreen
 
         [InputControl(layout = "Button")]
         [SerializeField]
-        private string m_ControlPath;
+        internal string m_ControlPath;
 
         protected override string controlPathInternal
         {
             get => m_ControlPath;
             set => m_ControlPath = value;
         }
-
-#if UNITY_EDITOR
-        [UnityEditor.CustomEditor(typeof(OnScreenButton))]
-        internal class OnScreenButtonEditor : UnityEditor.Editor
-        {
-            private UnityEditor.SerializedProperty m_ControlPathInternal;
-
-            public void OnEnable()
-            {
-                m_ControlPathInternal = serializedObject.FindProperty(nameof(OnScreenButton.m_ControlPath));
-            }
-
-            public void OnDisable()
-            {
-                new InputComponentEditorAnalytic(InputSystemComponent.OnScreenButton).Send();
-            }
-
-            public override void OnInspectorGUI()
-            {
-                // Current implementation has UGUI dependencies (ISXB-915, ISXB-916)
-                UGUIOnScreenControlEditorUtils.ShowWarningIfNotPartOfCanvasHierarchy((OnScreenButton)target);
-
-                UnityEditor.EditorGUILayout.PropertyField(m_ControlPathInternal);
-
-                serializedObject.ApplyModifiedProperties();
-            }
-        }
-#endif
     }
 }
 #endif
