+++ Packages/com.unity.inputsystem/Runtime/Plugins/PlayerInput/PlayerInput.cs
@@ -6,10 +6,6 @@ using UnityEngine.InputSystem.Users;
 using UnityEngine.InputSystem.Utilities;
 using UnityEngine.InputSystem.OnScreen;
 
-#if UNITY_EDITOR
-using UnityEngine.InputSystem.Editor;
-#endif
-
 #if PACKAGE_DOCS_GENERATION || UNITY_INPUT_SYSTEM_ENABLE_UI
 using UnityEngine.InputSystem.UI;
 #endif
